Move tests pynini to integration tests
